+++ sparkle/write.py
@@ -5,22 +5,23 @@ try:
 except ImportError:
     from urlparse import urlparse
 
-from sparkle.utils import context_has_package, config_reader_writer, context_has_jar
+from sparkle.utils import (context_has_package, config_reader_writer,
+                           context_has_jar, to_parsed_url_and_options)
 
 
-def fs(df, path, partition_by=None, mode=None, format=None, options=None):
-    """Writes dataframe to s3.
+def fs(df, path, partition_by=None, mode=None, output_format=None, options=None):
+    """Writes dataframe to a file system.
 
     Args:
         df (pyspark.sql.DataFrame):
-        path (str): s3 or local fs path.
-        partition_by (list): fields to partition by.
-        format (str|None): output files format.
+        path (str): S3 or local fs path.
+        partition_by (list): Fields to partition by.
+        output_format (str|None): Output files format.
     """
     writer = df.write\
-        .mode(mode or 'overwrite')\
+        .mode(mode)\
         .partitionBy(*partition_by)\
-        .format(format or 'parquet')
+        .format(output_format)
     config_reader_writer(writer, options).save(path)
 
 
@@ -32,7 +33,7 @@ def cassandra(df, host, keyspace, table, consistency=None, mode=None, options=No
         host (str)
         keyspace (str)
         table (str)
-        consistency (str|None): write consitency level: ONE, QUORUM, ALL, etc.
+        consistency (str|None): Write consitency level: ONE, QUORUM, ALL, etc.
         mode (str|None): Spark save mode,
             http://spark.apache.org/docs/latest/sql-programming-guide.html#save-modes
         options (dict[str, str]): Additional options to `org.apache.spark.sql.cassandra` format.
@@ -41,15 +42,15 @@ def cassandra(df, host, keyspace, table, consistency=None, mode=None, options=No
 
     writer = config_reader_writer(df.write.format('org.apache.spark.sql.cassandra'), {
         'spark_cassandra_connection_host': host,
-        'spark_cassandra_output_consistency_level': consistency or 'QUORUM',
+        'spark_cassandra_output_consistency_level': consistency,
         'keyspace': keyspace,
         'table': table,
     })
 
-    config_reader_writer(writer, options).mode(mode or 'append').save()
+    config_reader_writer(writer, options).mode(mode).save()
 
 
-def csv(df, path, header=False, mode='error', options=None):
+def csv(df, path, header=False, mode=None, options=None):
     """Write dataframe into csv file.
 
     Args:
@@ -86,7 +87,7 @@ def elastic(df, host, es_index, es_type, mode=None, options=None):
     })
 
     config_reader_writer(writer, options).\
-        mode(mode or 'append').\
+        mode(mode).\
         save('{}/{}'.format(es_index, es_type))
 
 
@@ -97,9 +98,9 @@ def mysql(df, host, database, table, mode=None, options=None):
         df (pyspark.sql.DataFrame)
         host (str)
         database (str)
-        table (str): mysql table
-        mode (str|None): write mode
-        options (dict): additional options
+        table (str): Mysql table.
+        mode (str|None): Write mode.
+        options (dict): Additional options.
     """
     assert context_has_jar(df.sql_ctx, 'mysql-connector-java')
 
@@ -109,7 +110,7 @@ def mysql(df, host, database, table, mode=None, options=None):
     df.write.jdbc(
         'jdbc:mysql://{}:3306/{}'.format(host, database),
         table,
-        mode=mode or 'append',
+        mode=mode,
         properties=options,
     )
 
@@ -137,6 +138,13 @@ def by_url(df, url):
         df (pyspark.sql.DataFrame): data to be written.
         url (str): url of destination.
     """
+    _by_url_registry = {
+        'parquet': _fs_resolver,
+        'csv': _fs_resolver,
+        'cassandra': _cassandra_resolver,
+        'mysql': _mysql_resolver,
+        'elastic': _elastic_resolver,
+    }
     scheme = urlparse(url).scheme
     try:
         _by_url_registry[scheme](df, url)
@@ -145,22 +153,21 @@ def by_url(df, url):
 
 
 def _fs_resolver(df, url):
-    inp, options = _to_parsed_url_and_options(url)
-
+    inp, options = to_parsed_url_and_options(url)
     partition_by = options.pop('partition_by', None)
-    format_ = inp.scheme
+    output_format = inp.scheme
     mode = options.pop('mode', None)
     fs(df,
        inp.path,
        partition_by=partition_by.split(',') if partition_by else None,
-       format=format_,
+       output_format=output_format,
        mode=mode,
        options=options,
        )
 
 
 def _cassandra_resolver(df, url):
-    inp, options = _to_parsed_url_and_options(url)
+    inp, options = to_parsed_url_and_options(url)
     _, db, table = inp.path.split('/')
     mode = options.pop('mode', None)
     consistency = options.pop('consistency', None)
@@ -175,7 +182,7 @@ def _cassandra_resolver(df, url):
 
 
 def _mysql_resolver(df, url):
-    inp, options = _to_parsed_url_and_options(url)
+    inp, options = to_parsed_url_and_options(url)
     _, db, table = inp.path.split('/')
     mode = options.pop('mode', None)
     mysql(
@@ -188,7 +195,7 @@ def _mysql_resolver(df, url):
 
 
 def _elastic_resolver(df, url):
-    inp, options = _to_parsed_url_and_options(url)
+    inp, options = to_parsed_url_and_options(url)
     _, index, type_ = inp.path.split('/')
     mode = options.pop('mode', None)
     elastic(
@@ -198,29 +205,3 @@ def _elastic_resolver(df, url):
         mode=mode,
         options=options,
     )
-
-
-_by_url_registry = {
-    # scheme: resolver
-}
-
-
-def _to_parsed_url_and_options(url):
-    """Returns parsed url and options."""
-    inp = urlparse(url)
-    options = dict([item.split('=', 1) for item in inp.query.split('&')]) if inp.query else {}
-    return inp, options
-
-
-def _register_by_resolver(scheme, resolver):
-    if scheme in _by_url_registry:
-        raise Exception('The scheme {} already have a resolver'.format(scheme))
-
-    _by_url_registry[scheme] = resolver
-
-
-_register_by_resolver('parquet', _fs_resolver)
-_register_by_resolver('csv', _fs_resolver)
-_register_by_resolver('cassandra', _cassandra_resolver)
-_register_by_resolver('mysql', _mysql_resolver)
-_register_by_resolver('elastic', _elastic_resolver)